# plaid-service/api/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Backbase Plaid Integration Service
  license:
    name: Backbase
servers:
  - url:
paths:
  /links/request_plaid_link:
    post:
      summary: Request plaid link
      operationId: requestPlaidLink
      tags:
        - link
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaidLinkRequest'
      responses:
        '200':
          description: Plaid Link to use in when invoking the Plaid Link UI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaidLinkResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /links/set_access_token:
    post:
      summary: Set Public Access Token
      operationId: setPublicAccessToken
      tags:
        - link
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAccessTokenRequest'
      responses:
        '201':
          description: A paged array of pets
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /item/{itemId}:
    delete:
      tags:
        - item
      summary: "Delete item"
      operationId: "deleteItem"
      parameters:
        - name: itemId
          description: The plaid item id to delete
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: "Deleted item complete"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /webhook/{itemId}:
    post:
      summary: Allow PLAID to push updates into DBS
      operationId: processWebHook
      parameters:
        - name: itemId
          description: The plaid item id to apply the webhook to
          in: path
          required: true
          schema:
            type: string
      tags:
        - webhook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaidWebhook'
      responses:
        '202':
          description: "Accepted webhook"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /webhook/{itemId}/refresh-transactions:
    post:
      summary: Force refresh of transactions for item
      operationId: refreshTransactions
      parameters:
        - name: itemId
          description: The plaid item id to refresh
          in: path
          required: true
          schema:
            type: string
      tags:
        - webhook
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                all:
                  type: boolean
                  default: true
      responses:
        '202':
          description: "Accepted webhook"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    PlaidLinkRequest:
      type: object
      properties:
        language:
          type: string
        name:
          type: string
    PublicTokenRequest:
      type: object
      properties:
        public-token:
          type: string
      required:
        - public-token
    PlaidLinkResponse:
      type: object
      properties:
        token:
          type: string
    SetAccessTokenRequest:
      properties:
        public_token:
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'
    PlaidItem:
      type: object
      properties:
        available_products:
          type: array
          items:
            type: string
        billed_products:
          type: array
          items:
            type: string
        consent_expiration_time:
          type: string
        error:
          $ref: '#/components/schemas/Error'
        institution_id:
          type: string
        item_id:
          type: string
        webhook:
          type: string
    PlaidWebhook:
      type: object
      properties:
        webhook_type:
          type: string
          enum:
            - TRANSACTIONS
            - ITEM
        webhook_code:
          type: string
          enum:
            - INITIAL_UPDATE
            - HISTORICAL_UPDATE
            - DEFAULT_UPDATE
            - TRANSACTIONS_REMOVED
            - WEBHOOK_UPDATE_ACKNOWLEDGED
            - ERROR
        item_id:
          type: string
        error:
          type: string
        new_transactions:
          type: integer
        removed_transactions:
          type: array
          items:
            type: string
    Metadata:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/PlaidAccount'
        account_id:
          type: string
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/PlaidAccount'
        institution:
          $ref: '#/components/schemas/PlaidInstitution'

    PlaidAccount:
      type: object
      properties:
        id:
          type: string
        mask:
          type: string
        name:
          type: string
        subtype:
          type: string
        type:
          type: string
    PlaidInstitution:
      type: object
      properties:
        institution_id:
          type: string
        name:
          type: string
    PlaidTransaction:
      type: object
      properties:
        account_id:
          type: string
        amount:
          type: number
        iso_currency_code:
          type: string
        category:
          type: array
          items:
            type: string
        category_id:
          type: string
        date:
          type: string
          format: date
        authorized_date:
          type: string
          format: date
        location:
          $ref: '#/components/schemas/Location'
        name:
          type: string
        merchant_name:
          type: string
        payment_meta:
          $ref: '#/components/schemas/Payment_meta'
        payment_channel:
          type: string
        pending:
          type: boolean
        pending_transaction_id:
          type: string
        account_owner:
          type: string
        transaction_code:
          type: string
        transaction_type:
          type: string
    Payment_meta:
      type: object
      properties:
        by_order_of:
          type: string
        payee:
          type: string
        payer:
          type: string
        payment_method:
          type: string
        payment_processor:
          type: string
        ppd_id:
          type: string
        reason:
          type: string
        reference:
          type: string
    Location:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        region:
          type: string
        postal_code:
          type: string
        country:
          type: string
        lat:
          type: number
        lon:
          type: number
        store_number:
          type: integer

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string